Fix image save (#13)

+++ tests/cases/ImageStorageTest.php
@@ -3,6 +3,9 @@
 namespace Tests\Cases;
 
 use Contributte\ImageStorage\ImageStorage;
+use Exception;
+use Nette\Http\FileUpload;
+use Nette\Utils\Image;
 use Ninjify\Nunjuck\TestCase\BaseTestCase;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
@@ -31,15 +34,19 @@ final class ImageStorageTest extends BaseTestCase
 	}
 
 
-	public static function recursiveRemove(string $path): void
+	public function tearDown(): void
 	{
-		$iterator = new RecursiveDirectoryIterator($path, RecursiveDirectoryIterator::SKIP_DOTS);
-		$files = new RecursiveIteratorIterator($iterator, RecursiveIteratorIterator::CHILD_FIRST);
-		foreach ($files as $file) {
-			if ($file->isDir()) {
-				rmdir($file->getRealPath());
-			} else {
-				unlink($file->getRealPath());
+		$path = __DIR__ . '/../data/images';
+
+		if (file_exists($path)) {
+			$iterator = new RecursiveDirectoryIterator($path, RecursiveDirectoryIterator::SKIP_DOTS);
+			$files = new RecursiveIteratorIterator($iterator, RecursiveIteratorIterator::CHILD_FIRST);
+			foreach ($files as $file) {
+				if ($file->isDir()) {
+					rmdir($file->getRealPath());
+				} else {
+					unlink($file->getRealPath());
+				}
 			}
 		}
 	}
@@ -73,6 +80,82 @@ final class ImageStorageTest extends BaseTestCase
 
 		foreach ($file_array as $name) {
 			Assert::falsey(file_exists($name));
+		}
+	}
+
+
+	public function testUpload(): void
+	{
+		$files = __DIR__ . '/../data/files';
+		$tempImagePath = $files . '/tmp.jpg';
+
+		$imageContent = Image::fromBlank(1, 1)->toString();
+
+		$this->saveTempImage($tempImagePath, $imageContent);
+
+		$upload = new FileUpload([
+			'name' => 'upload.jpg',
+			'type' => 'image/jpg',
+			'size' => '20',
+			'tmp_name' => $tempImagePath,
+			'error' => 0,
+		]);
+
+		$this->storage->saveUpload($upload, 'images');
+
+		$prefix = $this->getPrefixFromContent($imageContent);
+
+		$savedImage = sprintf(
+			'%s/../images/%s/upload.jpg',
+			$files,
+			$prefix
+		);
+		Assert::truthy(file_exists($savedImage));
+	}
+
+
+	public function testSaveContent(): void
+	{
+		$imageFileName = 'content.jpg';
+		$files = __DIR__ . '/../data/files';
+
+		$imageContent = Image::fromBlank(1, 1)->toString();
+
+		$prefix = $this->getPrefixFromContent($imageContent);
+
+		$this->storage->saveContent($imageContent, $imageFileName, 'images');
+		$savedImage = sprintf(
+			'%s/../images/%s/content.jpg',
+			$files,
+			$prefix
+		);
+		Assert::truthy(file_exists($savedImage));
+
+		$this->storage->saveContent($imageContent, $imageFileName, 'images');
+		$savedImageCopy = sprintf(
+			'%s/../images/%s/content.2.jpg',
+			$files,
+			$prefix
+		);
+		Assert::truthy(file_exists($savedImageCopy));
+	}
+
+
+	private function getPrefixFromContent(string $imageContent): string
+	{
+		return substr(sha1($imageContent), 0, 2);
+	}
+
+
+	/**
+	 * @throws Exception
+	 */
+	private function saveTempImage(string $path, string $content): void
+	{
+		$result = file_put_contents($path, $content);
+
+		if ($result === false) {
+			throw new Exception('Unable to save temporary test image!');
 		}
 	}
